feat(test): add shared MongoDB test container for integration tests
